Update build to run on Java 21
We specify an arbitrary version to build Gradle with in order to ensure build script compilation
hits the cache across machines -- where unless restricted in this manner -- would allow developers
to use different Java versions for the build.

However, we do verify the build succeeds with various Java versions, up to 21, with the Gradleception builds.

This commit updates the build to require Java 21.