Introduce API to request feature of dependency
This commit introduces the concept of a CapabilitySelector, with two implementations,
the ExactCapabilitySelector and the FeatureCapabilitySelector.

This concept is similar to that of a ComponentSelector, where we have some criteria
for the capability we are looking for, but do not specify the exact capability we want.

A FeatureCapabilitySelector can be specified with only a feature name. For test fixtures,
the feature name would be test-fixtures. Then, once a target component has been selected
from a component selector, we can use the target component's module coordinates to
derive a complete capability coordinate set from the feature capability selector.

This allows us to request capabilities of a target component without knowing the
coordinates of the target component. Further, this plays nicely with dependency
substitutions, where it can be difficult or impossible to substitute a dependency with another
when using capabilities, as the coordinates of the capability would change along with the
substituted dependency. Using feature component selectors, only the feature name is carried,
and substutions are transparent.

The ExactCapabilitySelector allows the user to request a group/name outright, as the current API does.

In practice, a Capability only belongs on the producer side, much like a ComponentIdentifier only belongs
on the producer component. A CapabilitySelector and ComponentSelector belong on the consumer side.
In fact, the existing API for requesting capabilities includes the possiblility of specifying a version,
however when resolving the capability against a target component, we always ignored the version.
Because of this, we did not include a version field in the ExactCapabilitySelector definition.

There are some slight behavior changes in this commit. Mostly along the lines of printing capabilities
without their versions when they are used as a capability selector. This is because we do not carry the
version, as it is ignored when performing selection.

Furthermore, when publishing requested capabilities, we also no longer publish the version, as it will
be ignored when consuming the metadata. Feature capability selectors are 'resolved' when published,
to their exact counterparts. In the future we may want to improve GMM to have 'feature selector' support,
however this commit does not make that enhancement.

Finally, the true purpose of this commit is to introduce an Isolated Projects-safe mechanism for requesting
the test fixtures of a target project component. For this reason, we have also migrated test fixtures to
build on top of this feature.